Port Ycash per-Sapling-output anti-spam fee floor; ZIP 317 off by default
Restores the current Ycash code mempool anti-spam policy: a
per-Sapling-output fee floor enforced in AcceptToMemoryPool
(1000 yoshi within a 50-output grace zone, then 1000 yoshi
per additional output). Uses GetSaplingOutputsCount() so v4
and v5 transactions both work.

ZIP 317 unpaid-action limits are neutralized by default
(DEFAULT_TX_/BLOCK_UNPAID_ACTION_LIMIT = SIZE_MAX); all
ZIP 317/401 code stays compiled and exercised. Per-node
opt-in via -txunpaidactionlimit=0 -blockunpaidactionlimit=0;
a network-wide flip is a one-line default change (relay
policy, not consensus). Activation path documented in zip317.h.

Fix nTxUnpaidActionLimit type (CAmount -> size_t): SIZE_MAX
assigned to int64_t wraps to -1 and rejects every tx.

Bundled cleanups: TRANSPARENT_VALUE_VERSION bumped to 6120350
(nChainSupplyDelta tracking starts now); NU6_DATA_VERSION
dropped (unused); ZIP212_GRACE_PERIOD 32256 -> 3 to match
current Ycash code (upstream value had drifted back during a
librustzcash sync); UPGRADE_NU6/NU6_1 comments mark them as
inert placeholders on Ycash.

+++ src/init.cpp
@@ -1241,7 +1241,7 @@ bool AppInit2(boost::thread_group& threadGroup, CScheduler& scheduler)
     if (GetArg("-txunpaidactionlimit", 0) < 0) {
         return InitError(_("-txunpaidactionlimit cannot be configured with a negative value."));
     }
-    nTxUnpaidActionLimit = GetArg("-txunpaidactionlimit", DEFAULT_TX_UNPAID_ACTION_LIMIT);
+    nTxUnpaidActionLimit = (size_t) GetArg("-txunpaidactionlimit", (int64_t) DEFAULT_TX_UNPAID_ACTION_LIMIT);
 
     // Option to startup with mocktime set (used for regression testing);
     // a mocktime of 0 (the default) selects the system clock.
